refactor: move utils hooks outside pages
