+++ examples/realtime-console/src/hooks/use-access-token.tsx
