+++ examples/realtime-console/src/utils/constants.ts
