+++ examples/realtime-console/src/utils/local-manager.ts
